+++ src/core/codecs/lucene90/lucene90_points_writer.rs
@@ -202,8 +202,9 @@ where
         CodecUtil::write_footer(&mut self.index_out)?;
         CodecUtil::write_footer(&mut self.data_out)?;
         self.meta_out
-            .write_long(self.index_out.get_file_pointer())?;
-        self.meta_out.write_long(self.data_out.get_file_pointer())?;
+            .write_long(self.index_out.get_file_pointer() as i64)?;
+        self.meta_out
+            .write_long(self.data_out.get_file_pointer() as i64)?;
         CodecUtil::write_footer(&mut self.meta_out)?;
         Ok(())
     }
