+++ src/core/store/buffered_index_input_base.rs
@@ -27,20 +27,24 @@ pub trait BufferedIndexInputBase: crate::core::util::clone::TryClone {
     ///
     /// # See Also
     /// [`read_internal`](BufferedIndexInputBase::read_internal)
-    fn seek_internal(&mut self, pos: i64) -> Result<()>;
+    fn seek_internal(&mut self, pos: usize) -> Result<()>;
     /// Expert: Implements buffer refill. Reads bytes from the current position
     /// in the input.
     ///
     /// # Arguments
     /// * `b` - The buffer to read bytes into.
-    fn read_internal(&mut self, b: &mut Cursor<Vec<u8>>, len: i64, file_pointer: i64)
-    -> Result<()>;
+    fn read_internal(
+        &mut self,
+        b: &mut Cursor<Vec<u8>>,
+        len: usize,
+        file_pointer: usize,
+    ) -> Result<()>;
 
     /// Creates a slice of this index input, with the given description, offset,
     /// and length. The slice is positioned at the beginning.
     type Slice: IndexInput + RandomAccessInput;
-    fn slice(&self, slice_description: &str, offset: i64, length: i64) -> Result<Self::Slice>;
+    fn slice(&self, slice_description: &str, offset: usize, length: usize) -> Result<Self::Slice>;
 
     /// The number of bytes in the file.
-    fn length(&self) -> i64;
+    fn length(&self) -> usize;
 }
