+++ src/core/store/dummy/dummy_index_output.rs
@@ -27,7 +27,7 @@ impl DataOutput for DummyIndexOutput {
         unreachable!("Dummy implementation: this method should never be called in real usage")
     }
 
-    fn write_bytes_range(&mut self, _b: &[u8], _offset: i32, _length: i32) -> Result<()> {
+    fn write_bytes_range(&mut self, _b: &[u8], _offset: usize, _length: usize) -> Result<()> {
         unreachable!("Dummy implementation: this method should never be called in real usage")
     }
 }
@@ -45,7 +45,7 @@ impl Closeable for DummyIndexOutput {
 }
 
 impl IndexOutput for DummyIndexOutput {
-    fn get_file_pointer(&self) -> i64 {
+    fn get_file_pointer(&self) -> usize {
         unreachable!("Dummy implementation: this method should never be called in real usage")
     }
